[MISC] move type_traits/deferred_crtp_base to detail
