+++ include/seqan3/alignment/pairwise/alignment_algorithm.hpp
@@ -12,8 +12,10 @@
 
 #pragma once
 
+#include <seqan3/std/iterator>
 #include <memory>
 #include <optional>
+#include <seqan3/std/ranges>
 #include <type_traits>
 
 #include <seqan3/alignment/configuration/align_config_band.hpp>
@@ -24,20 +26,17 @@
 #include <seqan3/alignment/pairwise/detail/concept.hpp>
 #include <seqan3/alignment/pairwise/detail/type_traits.hpp>
 #include <seqan3/alignment/matrix/detail/aligned_sequence_builder.hpp>
-
+#include <seqan3/core/detail/deferred_crtp_base.hpp>
 #include <seqan3/core/detail/empty_type.hpp>
 #include <seqan3/core/simd/concept.hpp>
 #include <seqan3/core/simd/simd.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
 #include <seqan3/core/simd/view_to_simd.hpp>
-#include <seqan3/core/type_traits/deferred_crtp_base.hpp>
 #include <seqan3/core/type_traits/function.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 #include <seqan3/range/views/drop.hpp>
 #include <seqan3/range/views/get.hpp>
 #include <seqan3/range/views/take.hpp>
-#include <seqan3/std/iterator>
-#include <seqan3/std/ranges>
 
 namespace seqan3::detail
 {
