+++ include/seqan3/core/type_traits/all.hpp
@@ -17,8 +17,8 @@
 
 #pragma once
 
+#include <seqan3/core/detail/deferred_crtp_base.hpp>
 #include <seqan3/core/type_traits/pre.hpp>
-#include <seqan3/core/type_traits/deferred_crtp_base.hpp>
 #include <seqan3/core/type_traits/iterator.hpp>
 #include <seqan3/core/type_traits/range.hpp>
 #include <seqan3/core/type_traits/template_inspection.hpp>
