+++ test/snippet/core/type_traits/deferred_crtp_base.cpp
@@ -1,7 +1,7 @@
 #include <string>
 #include <type_traits>
 
-#include <seqan3/core/type_traits/deferred_crtp_base.hpp>
+#include <seqan3/core/detail/deferred_crtp_base.hpp>
 
 // Defines a crtp_base class with an additional value type.
 template <typename derived_t, int value>
